Backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.0
Flask-Limiter==3.8.0
Flask-Talisman==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==23.0.0
bcrypt==4.2.1
marshmallow==3.23.1
bleach==6.2.0